Refactor: Extract main logic into pure functions and separate modules
